Organize and document extension code
