+++ package.json
@@ -22,7 +22,7 @@
     "pop-observe": "^2.0.1",
     "pop-swap": "^1.0.0",
     "q": "^2.0.3",
-    "system": "^1.1.1",
+    "system": "^1.2.1",
     "wizdom": "^1.1.0"
   },
   "devDependencies": {
