--- translate-xml.js
@@ -1,8 +1,0 @@
-"use strict";
-
-var translate = require("./translate");
-
-module.exports = function translateHtml(module) {
-    return translate(module, "application/xml");
-};
-
